Remove redundant NOPs during bytecode optimization.
NOP instruction is required iff one of the following is true:
(a) it is a sole bytecode instruction in a try-catch block (empty TCBs are not allowed);
(b) it is a sole bytecode instruction in a source code line (breakpoints on that line will not work).
All other NOP instructions can be removed.

Note that it doesn't really affect the performance for mature JVM implementations.
However, the perceived quality of the generated code is somewhat improved :).

Related: KT-15609