Move TypedArray's and is-checks for PrimitiveArray's behind -Xtypedarrays
Added a compilation key '-Xtypedarrays' which enables translation to
TypedArray's and is-checks for primitive arrays.

Added box tests for codegen/box/arrays with the key enabled.

Rewrote the primitiveArrays.kt test so that it works with the
key disabled,

Tweaked runtime, so that it works when compiled with the flag
disabled.